auth/requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# HTTP客户端
httpx>=0.25.0
requests>=2.31.0

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 中间件和工具
starlette>=0.27.0
python-dateutil>=2.8.2 

jinja2